スクリプト分離
