+++ __init__.py
@@ -34,6 +34,8 @@ imports = [
     "translations",
     "operator_separate_shapekeys",
     "operator_apply_modifiers",
+    "operator_separate_lr_shapekey_all",
+    "operator_separate_lr_shapekey_all_tag_detect",
     "menu_object_context",
     "menu_edit_mesh_context",
     "link_with_AutoMerge",
