+++ scripts/ShapeKeysUtil.py
@@ -21,42 +21,6 @@ from bpy.props import BoolProperty, EnumProperty, FloatProperty, IntProperty, Fl
 from . import func_utils, consts, func_separate_lr_shapekey, func_select_axis_from_point
 
 
-def separate_lr_shapekey_all(duplicate, enable_sort, auto_detect):
-    obj = func_utils.get_active_object()
-    print("Create LR Shapekey All: ["+obj.name+"]")
-    
-    # 頂点を全て表示
-    bpy.ops.object.mode_set(mode='EDIT')
-    bpy.ops.mesh.reveal()
-    bpy.ops.object.mode_set(mode='OBJECT')
-    
-    shape_keys_length=len(obj.data.shape_keys.key_blocks)
-    for i in reversed(range(shape_keys_length)):
-        if i == 0:
-            break
-        shapekey = obj.data.shape_keys.key_blocks[i]
-
-        # 名前の最後が_leftまたは_rightのシェイプキーは既に左右分割済みと見なし処理スキップ
-        if shapekey.name.endswith("_left") or shapekey.name.endswith("_right"):
-            continue
-        # auto_detectがTrueなら、名前に"%LR%"を含むときだけ左右分割処理を行う
-        if not auto_detect or (auto_detect and shapekey.name.find(consts.ENABLE_LR_TAG) != -1):
-            # print("Shapekey: ["+shapekey.name+"] ["+str(shape_keys_length-1-i)+" / "+str(shape_keys_length)+"]")
-            print("Shapekey: ["+shapekey.name+"]")
-            dup_temp = duplicate
-            sort_temp = enable_sort
-            if auto_detect:
-                # 名前に"%DUP%"を含むなら強制的に複製ON
-                if shapekey.name.find(consts.ENABLE_DUPLICATE_TAG) != -1:
-                    dup_temp = True
-                # 名前に"%SORT%"を含むなら強制的にソートON
-                if shapekey.name.find(consts.ENABLE_SORT_TAG) != -1:
-                    sort_temp = True
-            func_separate_lr_shapekey.separate_lr_shapekey(soruce_shape_key_index=i, duplicate=dup_temp, enable_sort=sort_temp)
-    
-    print("Finish Create LR Shapekey All: ["+obj.name+"]")
-
-
 # AddonPreferences #
 class addon_preferences(bpy.types.AddonPreferences):
     bl_idname = __package__
@@ -98,45 +62,6 @@ class OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey(bpy.types.Operator)
         bpy.ops.object.mode_set(mode='OBJECT')
         
         func_separate_lr_shapekey.separate_lr_shapekey(soruce_shape_key_index=obj.active_shape_key_index, duplicate=self.duplicate, enable_sort=self.enable_sort)
-        return {'FINISHED'}
-
-
-class OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey_all(bpy.types.Operator):
-    bl_idname = "object.shapekeys_util_separate_lr_shapekey_all"
-    bl_label = "Separate All Shape Key Left and Right"
-    bl_description = bpy.app.translations.pgettext(bl_idname+"_desc")
-    bl_options = {'REGISTER', 'UNDO'}
-    
-    duplicate: BoolProperty(name="Duplicate", default=False, description=bpy.app.translations.pgettext("separate_lr_shapekey_duplicate"))
-    enable_sort: BoolProperty(name="Enable Sort", default=False, description=bpy.app.translations.pgettext("separate_lr_shapekey_all_enable_sort"))
-    
-    @classmethod
-    def poll(cls, context):
-        obj = context.object
-        return obj.type == 'MESH' and obj.data.shape_keys and len(obj.data.shape_keys.key_blocks) != 0
-    
-    def execute(self, context):
-        obj = context.object
-        func_utils.set_active_object(obj)
-        separate_lr_shapekey_all(duplicate=self.duplicate, enable_sort=self.enable_sort, auto_detect=False)
-        return {'FINISHED'}
-
-
-class OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey_all_tagdetect(bpy.types.Operator):
-    bl_idname = "object.shapekeys_util_separate_lr_shapekey_all_tagdetect"
-    bl_label = "(Tag) Separate All Shape Key Left and Right"
-    bl_description = bpy.app.translations.pgettext(bl_idname+"_desc")
-    bl_options = {'REGISTER', 'UNDO'}
-    
-    @classmethod
-    def poll(cls, context):
-        obj = context.object
-        return obj.type == 'MESH' and obj.data.shape_keys and len(obj.data.shape_keys.key_blocks) != 0
-
-    def execute(self, context):
-        obj = context.object
-        func_utils.set_active_object(obj)
-        separate_lr_shapekey_all(duplicate=False, enable_sort=False, auto_detect=True)
         return {'FINISHED'}
 
 
@@ -263,8 +188,6 @@ class MESH_OT_specials_shapekeys_util_sideofactive_point(bpy.types.Operator):
 # Init #
 classes = [
     OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey,
-    OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey_all,
-    OBJECT_OT_specials_shapekeys_util_separate_lr_shapekey_all_tagdetect,
     OBJECT_OT_specials_shapekeys_util_assign_lr_shapekey_tag,
     
     MESH_OT_specials_shapekeys_util_sideofactive_point,
